// src/app/auth-service/Interfaces/iregister-tour-guide-data.ts
export interface IRegisterTourGuideData {
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  phoneNumber: string;
  address: string;
  guideName: string;
  licenseNumber: string;
  languages: string[];
  areasCovered: string;
  pricePerHour: number;
  contactEmail: string;
  contactPhone: string;
}